+++ tlUtils/config.js
@@ -5,3 +5,49 @@ export const GOERLI_RPC_URL = `https://goerli.infura.io/v3/2a4ce5a57b024046b1f2d
 
 // const MAINNET_RPC_URL = `https://mainnet.infura.io/v3/${infura}`
 // const GOERLI_RPC_URL = `https://goerli.infura.io/v3/${infura}`
+
+//WALLETS
+import infinityWalletModule from '@web3-onboard/infinity-wallet'
+import gnosisModule from '@web3-onboard/gnosis'
+import keepkeyModule from '@web3-onboard/keepkey'
+import keystoneModule from '@web3-onboard/keystone'
+import ledgerModule from '@web3-onboard/ledger'
+import sequenceModule from '@web3-onboard/sequence'
+import dcentModule from '@web3-onboard/dcent'
+import tahoModule from '@web3-onboard/sequence'
+import trustModule from '@web3-onboard/trust'
+import frontierModule from '@web3-onboard/frontier'
+import coinbaseWalletModule from '@web3-onboard/coinbase'
+import walletConnectModule from '@web3-onboard/walletconnect'
+
+import injectedModule from '@web3-onboard/injected-wallets'
+const injected = injectedModule()
+const coinbaseWalletSdk = coinbaseWalletModule()
+const walletConnect = walletConnectModule()
+const dcent = dcentModule()
+const infinityWallet = infinityWalletModule()
+const ledger = ledgerModule()
+const keystone = keystoneModule()
+const keepkey = keepkeyModule()
+const gnosis = gnosisModule()
+const sequence = sequenceModule()
+const taho = tahoModule() // Previously named Tally Ho wallet
+const trust = trustModule()
+const frontier = frontierModule()
+
+export const wallets = [
+    injected,
+    coinbaseWalletSdk,
+    walletConnect,
+    infinityWallet,
+    keepkey,
+    sequence,
+    injected,
+    trust,
+    frontier,
+    taho,
+    ledger,
+    dcent,
+    gnosis,
+    keystone,
+  ]
